refactor: move asset and link policy below editor
