+++ packages/core/service-core/src/workspace-asset-storage.ts
