+++ packages/shared/ws-path/src/index.ts
@@ -1,5 +1,6 @@
 export * from './asset-path';
 export * from './internal-link';
+export * from './markdown-link-target';
 export * from './route-strategy';
 export * from './workspace-file-policy';
 export * from './ws-path';
